requirements.txt
asgiref==3.2.10
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.1
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
idna==3.4
inflection==0.5.1
PyJWT==1.7.1
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
